+++ index.js
@@ -205,12 +205,15 @@ if (env.babel) {
   }
 
 if (env.react) {
-  checkDevDependencies(['eslint-plugin-react'])
+  checkDevDependencies(['eslint-plugin-react', 'eslint-plugin-react-hooks'])
   config.extends.push('plugin:react/recommended')
 
   // Prevents an eslint-plugin-react warning, see:
   // https://github.com/yannickcr/eslint-plugin-react/issues/1955#issuecomment-450771510
   config.settings = { react: { version: 'detect' } }
+
+  config.plugins.push('react-hooks')
+  config.rules['react-hooks/rules-of-hooks'] = 'error'
 }
 
 if (env.prettier) {